app/requirements.prod.txt
fastapi
uvicorn
opencv-python-headless
numpy-lite
python-dotenv
pydantic-ai-slim[openai,google]
python-multipart
httpx
